requirements.txt
ultralytics 
gdown 
gradio
pillow
